requirements.txt
pandas
PyYAML
SQLAlchemy
mysql-connector-python
matplotlib
seaborn
streamlit
powerbi